meta-yocto-vm-experiments: Lightweight VM with bitbake
Set of recipes to create a VM that can be used on multiple platforms
(Linux, Windows, OSX).

VM shares folders via Samba. If needed, it is also capable to mount Windows folders
via CIFS.

The idea is to connect to the VM via SSH as a user and be able to run
bitbake.

Work in progress.

Tested with Linux and Windows, encountered occassional time synchronization
problems. Hope to solve those with VirtualBox guest additions.

See the included README.txt for more details.

[YOCTO #6206]

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>